requirements.txt
# FastAPI 마이크로서비스 공통 종속성
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# MIDI 처리
mido==1.3.0
python-rtmidi==1.5.5

# RAG 및 AI
langchain==0.0.350
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2

# 모니터링 및 로깅
prometheus-client==0.19.0
structlog==23.2.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2